backend/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[silero,websocket,openai,deepgram]
loguru
google-auth-oauthlib
google-api-python-client
gspread-asyncio
openai